internal/runtime/maps: use uintptr instead of uint32 for index in group
This avoids some zero-extension ops on 64-bit machines.

Based on khr@'s CL 619479.

For #54766.

+++ src/internal/runtime/maps/table_debug.go
@@ -24,7 +24,7 @@ func (t *table) checkInvariants(m *Map) {
 	var empty uint64
 	for i := uint64(0); i <= t.groups.lengthMask; i++ {
 		g := t.groups.group(t.typ, i)
-		for j := uint32(0); j < abi.SwissMapGroupSlots; j++ {
+		for j := uintptr(0); j < abi.SwissMapGroupSlots; j++ {
 			c := g.ctrls().get(j)
 			switch {
 			case c == ctrlDeleted:
@@ -96,7 +96,7 @@ func (t *table) Print(m *Map) {
 
 		g := t.groups.group(m.typ, i)
 		ctrls := g.ctrls()
-		for j := uint32(0); j < abi.SwissMapGroupSlots; j++ {
+		for j := uintptr(0); j < abi.SwissMapGroupSlots; j++ {
 			print("\t\t\tslot ", j, "\n")
 
 			c := ctrls.get(j)
